# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/post/create": {
            "post": {
                "description": "Create post from body with authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Post"
                ],
                "summary": "Create post",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Auth token of logged-in user.",
                        "name": "X-Auth-Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "body params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/post.CreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpmodel.CreateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/post/delete/{post_id}": {
            "delete": {
                "description": "authenticates given user by giving an access jwttoken.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Post"
                ],
                "summary": "Delete post",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Auth token of logged-in user.",
                        "name": "X-Auth-Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID of the post to delete",
                        "name": "post_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/httpresponse.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/post/get/{post_id}": {
            "get": {
                "description": "authenticates given user by giving an access jwttoken.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Post"
                ],
                "summary": "Get post by id post",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Auth token of logged-in user.",
                        "name": "X-Auth-Token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/post.ReadPostResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/post/list": {
            "get": {
                "description": "authenticates given user by giving an access jwttoken.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Post"
                ],
                "summary": "List post",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Auth token of logged-in user.",
                        "name": "X-Auth-Token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/post.ReadPostResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/post/update": {
            "put": {
                "description": "Update post from body with authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Post"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Auth token of logged-in user.",
                        "name": "X-Auth-Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "body params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/post.UpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/httpresponse.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/post/update/{post_id}/image": {
            "put": {
                "description": "Update post image with authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Post"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Auth token of logged-in user.",
                        "name": "X-Auth-Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID of the post to update the image for",
                        "name": "post_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "The image file to upload",
                        "name": "image",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpmodel.UpdateImageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "httpmodel.CreateResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "ID of the created model`",
                    "type": "integer",
                    "x-order": "1",
                    "example": 1
                }
            }
        },
        "httpmodel.UpdateImageResponse": {
            "type": "object",
            "properties": {
                "uploaded_file_name": {
                    "description": "Uploaded file name",
                    "type": "string",
                    "x-order": "1",
                    "example": "https://res-cdn.com/postId"
                }
            }
        },
        "httpresponse.Response": {
            "type": "object",
            "properties": {
                "success": {
                    "type": "boolean",
                    "x-order": "1",
                    "example": true
                },
                "status_code": {
                    "type": "integer",
                    "x-order": "2",
                    "example": 200
                },
                "error": {
                    "x-order": "4",
                    "$ref": "#/definitions/httpresponse.ResponseError"
                },
                "data": {
                    "x-order": "5"
                }
            }
        },
        "httpresponse.ResponseError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "x-order": "1",
                    "example": "NOT_FOUND"
                },
                "detail": {
                    "type": "string",
                    "x-order": "2",
                    "example": "user not found"
                }
            }
        },
        "post.CreateRequest": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                }
            }
        },
        "post.ReadPostResponse": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image": {
                    "type": "string"
                }
            }
        },
        "post.UpdateRequest": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image": {
                    "type": "string"
                }
            }
        },
...
    }
}